requirements.txt
numpy==1.26.4
pandas==2.2.3
kagglehub==0.3.5
gradio==5.9.1
python-dotenv==1.0.1
langchain==0.3.12
langchain-chroma==0.1.4
langchain-community==0.3.12
langchain-openai==0.2.12
langchain-text-splitters==0.3.3
chromadb==0.5.23
transformers==4.47.1
torch==2.5.1
tqdm==4.67.1